requirements.txt
Flask==2.2.0
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-Cors==3.0.10
Flask-Reuploaded==1.2.0
gunicorn==20.1.0
pytest==7.0.1
python-dotenv==0.19.2
psycopg2-binary==2.9.3
Flask-Migrate==3.1.0
Werkzeug==2.2.0
coverage==6.4.4
PyPDF2
scholarly
openai
pyrebase
pycryptodome
Flask-WTF==1.1.1
WTForms==3.0.1
PyJWT==2.6.0